Fix disabling bug
This may fix the disabling issue. I didn't see the problem running the
lemonscript two ball auto at 800 Hz, or running drive straight at 800 Hz
for four minutes (tested three times). That being said, this should not
have fixed this, so I am not 100% sure that the issue is fixed.

The thing that seems to be correlated with the issue being fixed was
fixing the termination condition for driving straight. This may mean
that the issue exists in 971's drive code (When we're seeing the +12V on
one side and ~0V on the other). We should test this more, and see if the
termination bug was actually causing the issue before we move on with
this code.