Add AbsolutePointTurn and DriveDistanceAtAngle
AbsolutePointTurn will turn to a specific angle, relative to the
starting position of the robot. So, AbsolutePointTurn(0 * deg) will turn
to the starting point of the robot.

DriveDistanceAtAngle drives a distance, while setting the target gyro
angle to the angle passed to it relative to the starting gyro position.
This is to avoid drifting when driving forwards and backwards, and *NOT*
to be used when you may be off from the angle passed to it.

This commit also changes the vision code to turn to the angle passed to
it, regardless of the value of targetFound. This is most likely
needed due to a bug in the driver station vision code.

This commit also changes the termination condition of the motion
profiles: The angle must be within 0.5 degrees, and the distance within
2 cm.