Refactor vision
This splits the code in //vision into //vision/robot and
//vision/coprocessor, updating the BUILD files to reflect this. This
also changes the name of the vision targets to `main`.

In addition, this restructures the code in //vision/robot/vision.cpp and
//frc1678/frc1678.cpp so that it doesn't make my want to stab my eyes
out when I look at it. This includes:

* Making update a void that takes no arguments
* Making update call the function to check for alignment
* Making a getter for aligned_, so that the number of times that you
  call it doesn't matter

This code has not been tested, but *should* work in exactly the same
manner as the previous code. However, we may need to update the number
of required update cycles to be aligned for, as Aligned() was being
called many times in frc1678.cpp.